Updated node/pod HOSTNAME, HOST and IDENTITY and ADDRESS env variable with Kubernetes pod name from metadata name
This StatefulSet of Apache NiFi deployment uses hostname as reference in the cluster names. When Apache NiFi nodes/pods occasionally terminated and redeployed old zombie nodes/pods hostnames are still list in the cluster page and can not be removed, this will/should fix that bug

The HOSTNAME env variable is never empty as Kubernetes controller will always populate it if is not set

The NIFI_CLUSTER_NODE_ADDRESS requires FQDN else it will generate it own address and that will result in error.

Added setHostnameAsFQDN to Kubernetes StatefulSet deployment, this has been added as stable in Kubernetes 1.22 and we are currently running Kubernetes 1.27